python-opentelemetry-example/python-opentelemetry-example-fastapi/pyproject.toml
[project]
name = "python-opentelemetry-example-fastapi"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiohttp>=3.11.18",
    "fastapi[standard]>=0.115.12",
    "langchain>=0.3.25",
    "langchain-community>=0.3.23",
    "langchain-openai>=0.3.16",
    "langserve>=0.3.1",
    "opentelemetry-exporter-otlp>=1.32.1",
    "opentelemetry-instrumentation-aiohttp-client>=0.53b1",
    "opentelemetry-instrumentation-fastapi>=0.53b1",
    "opentelemetry-instrumentation-langchain>=0.40.2",
    "python-dotenv>=1.1.0",
    "sse-starlette>=2.3.3",
]

[dependency-groups]
dev = [
    "pyright>=1.1.400",
    "taskipy>=1.14.1",
]

[tool.taskipy.tasks]
fix = "uvx ruff format && uvx ruff check --fix && pyright"
